pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "laby"
dynamic = ["version"]
description = "Labyrinth Python SDK"
readme = "README.md"
authors = [{name = "James Sha", email = "james@laby.ai"}]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click",
    "PyYAML",
]

[project.scripts]
laby = "main:main"

[project.urls]
Homepage = "https://github.com/Labyrinth-Technologies/laby-py"
